SIG-1033 WIP split models.py
