+++ api/app/signals/apps/signals/models/buurt.py
@@ -1,0 +1,11 @@
+from django.contrib.gis.db import models
+
+
+class Buurt(models.Model):
+    ogc_fid = models.IntegerField(primary_key=True)
+    id = models.CharField(max_length=14)
+    vollcode = models.CharField(max_length=4)
+    naam = models.CharField(max_length=40)
+
+    class Meta:
+        db_table = 'buurt_simple'
