+++ api/app/signals/apps/signals/models/category.py
@@ -1,0 +1,72 @@
+from django.contrib.gis.db import models
+
+from django.utils.text import slugify
+
+
+class MainCategory(models.Model):
+    name = models.CharField(max_length=255, unique=True)
+    slug = models.SlugField(unique=True)
+
+    class Meta:
+        ordering = ('name',)
+        verbose_name_plural = 'Main Categories'
+
+    def __str__(self):
+        """String representation."""
+        return self.name
+
+    def save(self, *args, **kwargs):
+        self.slug = slugify(self.name)
+        super().save(*args, **kwargs)
+
+
+class Category(models.Model):
+    HANDLING_A3DMC = 'A3DMC'
+    HANDLING_A3DEC = 'A3DEC'
+    HANDLING_A3WMC = 'A3WMC'
+    HANDLING_A3WEC = 'A3WEC'
+    HANDLING_I5DMC = 'I5DMC'
+    HANDLING_STOPEC = 'STOPEC'
+    HANDLING_KLOKLICHTZC = 'KLOKLICHTZC'
+    HANDLING_GLADZC = 'GLADZC'
+    HANDLING_A3DEVOMC = 'A3DEVOMC'
+    HANDLING_WS1EC = 'WS1EC'
+    HANDLING_WS2EC = 'WS2EC'
+    HANDLING_REST = 'REST'
+    HANDLING_CHOICES = (
+        (HANDLING_A3DMC, HANDLING_A3DMC),
+        (HANDLING_A3DEC, HANDLING_A3DEC),
+        (HANDLING_A3WMC, HANDLING_A3WMC),
+        (HANDLING_A3WEC, HANDLING_A3WEC),
+        (HANDLING_I5DMC, HANDLING_I5DMC),
+        (HANDLING_STOPEC, HANDLING_STOPEC),
+        (HANDLING_KLOKLICHTZC, HANDLING_KLOKLICHTZC),
+        (HANDLING_GLADZC, HANDLING_GLADZC),
+        (HANDLING_A3DEVOMC, HANDLING_A3DEVOMC),
+        (HANDLING_WS1EC, HANDLING_WS1EC),
+        (HANDLING_WS2EC, HANDLING_WS2EC),
+        (HANDLING_REST, HANDLING_REST),
+    )
+
+    parent = models.ForeignKey('signals.MainCategory',
+                               related_name='categories',
+                               on_delete=models.PROTECT)
+    slug = models.SlugField()
+    name = models.CharField(max_length=255)
+    handling = models.CharField(max_length=20, choices=HANDLING_CHOICES)
+    departments = models.ManyToManyField('signals.Department')
+    is_active = models.BooleanField(default=True)
+
+    class Meta:
+        ordering = ('name',)
+        unique_together = ('parent', 'slug',)
+        verbose_name_plural = 'Sub Categories'
+
+    def __str__(self):
+        """String representation."""
+        return '{name} ({parent})'.format(name=self.name,
+                                          parent=self.parent.name)
+
+    def save(self, *args, **kwargs):
+        self.slug = slugify(self.name)
+        super().save(*args, **kwargs)
