+++ api/app/signals/apps/signals/models/department.py
@@ -1,0 +1,19 @@
+from django.contrib.gis.db import models
+
+
+class Department(models.Model):
+    code = models.CharField(max_length=3)
+    name = models.CharField(max_length=255)
+    is_intern = models.BooleanField(default=True)
+
+    class Meta:
+        ordering = ('name',)
+        indexes = [
+            models.Index(fields=['name']),
+        ]
+
+    def __str__(self):
+        """String representation."""
+        return '{code} ({name})'.format(code=self.code, name=self.name)
+
+
