+++ api/app/signals/apps/signals/models/mixins.py
@@ -1,0 +1,9 @@
+from django.contrib.gis.db import models
+
+
+class CreatedUpdatedModel(models.Model):
+    created_at = models.DateTimeField(editable=False, auto_now_add=True)
+    updated_at = models.DateTimeField(editable=False, auto_now=True)
+
+    class Meta:
+        abstract = True
