+++ api/app/signals/apps/signals/models/priority.py
@@ -1,0 +1,26 @@
+from django.contrib.gis.db import models
+
+from .mixins import CreatedUpdatedModel
+
+
+class Priority(CreatedUpdatedModel):
+    PRIORITY_NORMAL = 'normal'
+    PRIORITY_HIGH = 'high'
+    PRIORITY_CHOICES = (
+        (PRIORITY_NORMAL, 'Normal'),
+        (PRIORITY_HIGH, 'High'),
+    )
+
+    _signal = models.ForeignKey('signals.Signal',
+                                related_name='priorities',
+                                null=False,
+                                on_delete=models.CASCADE)
+    priority = models.CharField(max_length=10, choices=PRIORITY_CHOICES, default=PRIORITY_NORMAL)
+    created_by = models.EmailField(null=True, blank=True)
+
+    class Meta:
+        verbose_name_plural = 'Priorities'
+
+    def __str__(self):
+        """String representation."""
+        return self.get_priority_display()
