+++ api/app/signals/apps/signals/models/reporter.py
@@ -1,0 +1,19 @@
+from django.contrib.gis.db import models
+from django.contrib.postgres.fields import JSONField
+
+from .mixins import CreatedUpdatedModel
+
+
+class Reporter(CreatedUpdatedModel):
+    """Privacy sensitive information on reporter."""
+
+    _signal = models.ForeignKey(
+        "signals.Signal", related_name="reporters",
+        null=False, on_delete=models.CASCADE
+    )
+
+    email = models.EmailField(blank=True)
+    phone = models.CharField(max_length=17, blank=True)
+    remove_at = models.DateTimeField(null=True)
+
+    extra_properties = JSONField(null=True)  # TODO: candidate for removal
